% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{associationTest}
\alias{associationTest}
\title{Perform statistical test to check whether a gene is constant across
 a lineage}
\usage{
associationTest(models, global = TRUE, lineages = FALSE)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{global}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed. If the testing
 procedure was unsuccessful, the procedure will return NA test statistics and
 p-values.
}
\description{
Perform statistical test to check whether a gene is constant across
 a lineage
}
\examples{
data(gamList, package = "tradeSeq")
associationTest(gamList, global = TRUE, lineages = TRUE)
}
