% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cascade.R
\name{cascade}
\alias{cascade}
\alias{cascade,SingleCellExperiment-method}
\title{Get expression peak cascade of genes along a lineage.}
\usage{
cascade(models, ...)

\S4method{cascade}{SingleCellExperiment}(
  models,
  lineage = 1,
  genes = rownames(models),
  epsilon = 1e-06,
  nPoints = 100,
  derivativeThreshold = 0.1,
  derPvalThreshold = 0.05/nPoints,
  plotHeatmap = TRUE,
  clusterHeatmap = FALSE
)
}
\arguments{
\item{models}{The fitted GAMs. A \code{SingleCellExperiment} object as 
obtained from running \code{\link{fitGAM}}.}

\item{...}{parameters including:}

\item{lineage}{A positive integer, defining the lineage you would like to 
construct the cascade for. Defaults to \code{1}.}

\item{genes}{A character vector listing the genes from \code{models} for
which you would like to construct the cascade. Defaults to all genes present in \code{models}.}

\item{epsilon}{The distortion used for finite differencing. Defaults to \code{1e-6}.}

\item{nPoints}{The first derivatives are constructed pointwise on an equally-spaced
grid. The \code{nPoints} argument defines the number of points for grid construction.
Defaults to \code{100}.}

\item{derivativeThreshold}{The first derivative threshold tested against when
calculating test statistics for assessing significant increase. Defaults to \code{0.1}.}

\item{derPvalThreshold}{The p-value threshold used for determining significance of the
first derivative test statistic. Defaults to \code{0.05/nPoints}.}

\item{plotHeatmap}{Logical. Should a heatmap be plotted? Defaults to \code{TRUE}.}

\item{clusterHeatmap}{Logical. Should the heatmap cluster the genes? Defaults to \code{FALSE}.}
}
\value{
Only the genes with a significant expression increase will be returned.
For these genes, a list is returned
}
\description{
Finds and orders genes peaking in expression along a lineage.
The function first assesses which genes increase significantly in expression, somewhere
along a lineage, by testing pointwise first-derivatives against a threshold.
For genes found to be significantly increasing, we look for its expression peak.
}
\details{
First derivatives are estimated using finite differencing.
}
\examples{
set.seed(8)
data(crv, package="tradeSeq")
data(countMatrix, package="tradeSeq")
sce <- fitGAM(counts = as.matrix(countMatrix),
                  sds = crv,
                  nknots = 5)
# construct cascalde for lineage 1
cas <- cascade(sce)
}
