% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterExpressionPatterns.R
\docType{methods}
\name{clusterExpressionPatterns}
\alias{clusterExpressionPatterns}
\alias{.clusterExpressionPatterns}
\alias{clusterExpressionPatterns,SingleCellExperiment-method}
\alias{clusterExpressionPatterns,list-method}
\title{Cluster gene expression patterns.}
\usage{
.clusterExpressionPatterns(models, nPoints, genes, reduceMethod = "PCA",
  nReducedDims = 10, minSizes = 6, ncores = 1, random.seed = 176201,
  verbose = TRUE, ...)

\S4method{clusterExpressionPatterns}{SingleCellExperiment}(models, nPoints,
  genes, reduceMethod = "PCA", nReducedDims = 10, minSizes = 6,
  ncores = 1, random.seed = 176201, verbose = TRUE, ...)

\S4method{clusterExpressionPatterns}{list}(models, nPoints, genes,
  reduceMethod = "PCA", nReducedDims = 10, minSizes = 6, ncores = 1,
  random.seed = 176201, verbose = TRUE, ...)
}
\arguments{
\item{models}{The list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{nPoints}{The number of points to use for clustering the expression
patterns.}

\item{genes}{A numerical or character vector specifying the genes from
\code{models}
that should be clustered.}

\item{reduceMethod}{Method used before running the clustering methods.
Passed to \code{\link[clusterExperiment]{RSEC}}}

\item{nReducedDims}{Number of dimensions kept after \code{reduceMethod}.
Passed to \code{\link[clusterExperiment]{RSEC}}}

\item{minSizes}{Number of dimensions kept after \code{reduceMethod}.
Passed to \code{\link[clusterExperiment]{RSEC}}}

\item{ncores}{Number of cores to use. Passed to
\code{\link[clusterExperiment]{RSEC}}}

\item{random.seed}{Passed to \code{\link[clusterExperiment]{RSEC}}}

\item{verbose}{Passed to \code{\link[clusterExperiment]{RSEC}}}

\item{...}{Additional arguments to be passed to
\code{\link[clusterExperiment]{RSEC}}.}
}
\value{
A list containing the scaled fitted values \code{yhatScaled}(for
 plotting) and a \code{\link{ClusterExperiment}} object.
}
\details{
This method adopts the \code{\link[clusterExperiment]{RSEC}}
function from the clusterExperiment package to perform consensus clustering.
}
\examples{
data(gamList, package = "tradeSeq")
clusterExpressionPatterns(gamList, 200, seq_len(11))
}
