% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countMatrix}
\alias{countMatrix}
\title{A count matrix, used in the package vignette.}
\format{An object of class \code{dgCMatrix} with 240 rows and 2660 columns.}
\usage{
data(countMatrix)
}
\description{
This object contains the gene expression counts from the data described in
Paul et al. (2015).
}
\details{
#' @references
Franziska Paul, Yaara Arkin, Amir Giladi, DiegoAdhemar Jaitin, Ephraim Kenigsberg, Hadas KerenShaul, Deborah Winter, David Lara-Astiaso, Meital Gury, Assaf Weiner, Eyal David, Nadav Cohen,
FeliciaKathrineBratt Lauridsen, Simon Haas, Andreas Schlitzer, Alexander Mildner, Florent Ginhoux,
Steen Jung, Andreas Trumpp, BoTorben Porse, Amos Tanay, and Ido Amit. Transcriptional Heterogeneity and Lineage Commitment in Myeloid Progenitors. Cell, 163(7):1663–1677, 12 2015. ISSN 0092-
8674. doi: 10.1016/J.CELL.2015.11.013. URL https://www.sciencedirect.com/science/article/
ii/S0092867415014932?via%3Dihub#app3.
}
\keyword{datasets}
