% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{earlyDrivers}
\alias{earlyDrivers}
\title{Perform test of early differences between lineages}
\usage{
earlyDrivers(models)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform test of early differences between lineages
}
