% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{endPointOmnibusTest}
\alias{endPointOmnibusTest}
\title{Perform omnibus test to check for DE between final stages of every trajectory}
\usage{
endPointOmnibusTest(models, ...)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform omnibus test to check for DE between final stages of every trajectory
}
