% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_monocle_info}
\alias{extract_monocle_info}
\title{Extract info from Monocle models}
\usage{
extract_monocle_info(cds)
}
\arguments{
\item{cds}{A \code{\link{CellDataSet}}}
}
\value{
A list with four objects. A \code{pseudotime} matrix and a \code{cellWeights}
matrix that can be used as input to \code{\link{fitGAM}} or 
\code{\link{evaluateK}}, the reduced dimension matrix for the cells, and a 
list of length the number of lineages, containing the reduced dimension of 
each lineage.
}
\description{
This function extracts info that will be used downstream to make
\code{\link{CellDataSet}} objects compatible with a \code{\link{tradeSeq}}
analysis
}
\details{
For now, this only works for the DDRTree dimentionality reduction. 
It is the one recommanded by the Monocle developers.
}
