% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fitGAM}
\alias{fitGAM}
\title{Fit GAM model}
\usage{
fitGAM(counts, X = NULL, pseudotime, cellWeights, weights = NULL,
  seed = 81, offset = NULL)
}
\arguments{
\item{counts}{the count matrix.}

\item{X}{the design matrix of fixed effects. Defaults to only estimating
an intercept if not provided.}

\item{pseudotime}{a matrix of pseudotime values, each row represents a cell
and each column represents a lineage.}

\item{cellWeights}{a matrix of cell weights defining the probability that a
cell belongs to a particular lineage. Each row represents a cell and each
column represents a lineage.}

\item{weights}{a matrix of zero inflation weights with identical dimensions
as the \code{counts} matrix.}

\item{seed}{the seed used for assigning cells to trajectories}

\item{offset}{the offset, on log-scale, to account for differences in
sequencing depth.}
}
\description{
Fit GAM model
}
