% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamList}
\alias{gamList}
\title{A list of GAM models, used to demonstrate the various tests.}
\format{Can be re-obtained by runing the code in the example section
 of \code{\link{fitGAM}}.}
\usage{
gamList
}
\description{
A list of 11 \code{\link{gamObject}} obtained by fitting 10 genes on 15
cells randomly assigned to lineages with random pseudotimes.
}
\keyword{datasets}
