% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmootherPvalues.R
\name{getSmootherPvalues}
\alias{getSmootherPvalues}
\title{Get smoother p-value as returned by \code{mgcv}.}
\usage{
getSmootherPvalues(models)
}
\arguments{
\item{models}{the GAM models, typically the output from \code{\link{fitGAM}}.
Note that this function only works when \code{models} is a list.}
}
\value{
a matrix with the p-value associated with each lineage's smoother.
 The matrix has one row per gene where the fitting procedure converged.
}
\description{
Return smoother p-values from the \code{mgcv} package.
}
\examples{
data(gamList, package = "tradeSeq")
getSmootherPvalues(gamList)
}
