% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getSmootherTestStats}
\alias{getSmootherTestStats}
\title{Get smoother Chi-squared test statistics}
\usage{
getSmootherTestStats(models)
}
\arguments{
\item{models}{the GAM models, typically the output from \code{\link{fitGAM}}.}
}
\description{
Get smoother Chi-squared test statistics
}
