% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{identicalTest}
\alias{identicalTest}
\title{Perform test between lineages to check whether the gene dynamics are identical.}
\usage{
identicalTest(models)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{nPoints}{the number of points to be compared between lineages.}

\item{omnibus}{If TRUE, test for all pairwise comparisons simultaneously.}

\item{pairwise}{If TRUE, test for all pairwise comparisons independently.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value associated with each gene for all the tests performed.
}
\description{
Perform test between lineages to check whether the gene dynamics are identical.
}
\examples{
data(gamList, package = "tradeR")
identicalTest(gamList, omnibus = TRUE, pairwise = TRUE)
}
