% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nknots.R
\name{nknots}
\alias{nknots}
\alias{nknots,SingleCellExperiment-method}
\alias{nknots,list-method}
\title{knots}
\usage{
nknots(models, ...)

\S4method{nknots}{SingleCellExperiment}(models)

\S4method{nknots}{list}(models)
}
\arguments{
\item{models}{The fitted GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{...}{parameters including:}
}
\value{
A numeric, the number of nknots
}
\description{
Get the number of knots used for the fit
}
\examples{
data(gamList, package = "tradeSeq")
nknots(gamList)
}
