% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{patternTest}
\alias{patternTest}
\title{Perform pattern test between lineages}
\usage{
patternTest(models)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform pattern test between lineages
}
