% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCount.R
\name{plotGeneCount}
\alias{plotGeneCount}
\title{Plot gene expression in reduced dimension.}
\usage{
plotGeneCount(
  curve,
  counts = NULL,
  gene = NULL,
  clusters = NULL,
  models = NULL,
  title = NULL
)
}
\arguments{
\item{curve}{A \code{SlingshotDataSet} object. The output from trajectory inference
using Slingshot.}

\item{counts}{The count matrix, genes in rows and cells in columns.}

\item{gene}{The name of gene for which you want to plot the count or the row
number of that gene in the count matrix. Alternatively, one can specify
the \code{clusters} argument.}

\item{clusters}{The assignation of each cell to a cluster. Used to color the
plot. Either \code{clusters} or \code{gene} and \code{counts} must be supplied.}

\item{models}{The fitted GAMs, typically the output from
\code{\link{fitGAM}}. Used to display the knots.}

\item{title}{Title for the plot.}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot the gene in reduced dimensional space.
}
\details{
If both \code{gene} and \code{clusters} arguments are supplied, the
 plot will be colored according to gene count level.
}
\examples{
set.seed(97)
library(slingshot)
data(crv, package="tradeSeq")
data(countMatrix, package="tradeSeq")
rd <- slingshot::reducedDim(crv)
cl <- kmeans(rd, centers = 7)$cluster
lin <- slingshot::getLineages(rd, clusterLabels = cl, start.clus = 4)
crv <- slingshot::getCurves(lin)
counts <- as.matrix(countMatrix)
gamList <- fitGAM(counts = counts,
 pseudotime = slingPseudotime(crv, na = FALSE),
 cellWeights = slingCurveWeights(crv))
plotGeneCount(crv, counts, gene = "Mpo")
}
