% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSmoothers}
\alias{plotSmoothers}
\title{plot the model for a particular gene}
\usage{
plotSmoothers(m, nPoints = 100, ...)
}
\arguments{
\item{m}{the fitted model of a given gene}

\item{nPointss}{The number of points used to extraplolate the fit}
}
\description{
plot the model for a particular gene
}
