% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sds}
\alias{sds}
\title{A SlingshotDataset object, used in the package unit tests.}
\format{An object of class \code{SlingshotDataSet} of length 1.}
\source{
\url{https://bioconductor.org/packages/release/bioc/html/slingshot.html}
}
\usage{
data(sds)
}
\description{
This dataset contains the toy example from the Slingshot R package vignette.
}
\keyword{datasets}
