% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se}
\alias{se}
\title{The \code{\link{SingleCellExperiment}} object of bone marrow single-cell
 RNA-Seq data.}
\format{A \code{\link{SingleCellExperiment}} of 3004 genes and 2660 cells:
\describe{
See the help for \code{\link{SingleCellExperiment}} for more information.
}}
\source{
\url{https://cole-trapnell-lab.github.io/monocle-release/monocle3/}
}
\usage{
se
}
\description{
A dataset containing the scRNA-Seq count and UMAP reduced dimensions derived
 from following the Monocle 3 vignette.
}
\keyword{datasets}
