% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{startPointTest}
\alias{startPointTest}
\title{Perform statistical test to check for DE between starting point and the end
stages of every trajectory.}
\usage{
startPointTest(models, omnibus = TRUE, pairwise = FALSE, ...)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform statistical test to check for DE between starting point and the end
stages of every trajectory.
}
