% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{associationTest}
\alias{associationTest}
\title{Perform statistical test to check whether a gene is constant across a lineage}
\usage{
associationTest(models, omnibus = TRUE, lineages = FALSE, ...)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform statistical test to check whether a gene is constant across a lineage
}
