% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/associationTest.R
\name{associationTest}
\alias{associationTest}
\alias{associationTest,SingleCellExperiment-method}
\alias{associationTest,list-method}
\title{Perform statistical test to check whether gene expression is constant across
 pseudotime within a lineage}
\usage{
associationTest(models, ...)

\S4method{associationTest}{SingleCellExperiment}(models, global = TRUE, lineages = FALSE)

\S4method{associationTest}{list}(models, global = TRUE, lineages = FALSE)
}
\arguments{
\item{models}{The fitted GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{...}{parameters including:}

\item{global}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}
}
\value{
A matrix with the wald statistic, the number of
 degrees of freedom and the p-value
 associated with each gene for all the tests performed. If the testing
 procedure was unsuccessful, the procedure will return NA.
}
\description{
This test assesses whether average gene expression
is associated with pseudotime.
}
\examples{
data(gamList, package = "tradeSeq")
associationTest(gamList, global = TRUE, lineages = TRUE)
}
