% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{associationTest}
\alias{associationTest}
\title{Perform statistical test to check whether a gene is constant across a lineage}
\usage{
associationTest(models, omnibus = TRUE, lineages = FALSE)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{omnibus}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value associated with each gene for all the tests performed.
}
\description{
Perform statistical test to check whether a gene is constant across a lineage
}
\examples{
data(gamList, package = "tradeR")
associationTest(gamList, omnibus = TRUE, lineages = TRUE)
}
