% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{diffEndTest}
\alias{diffEndTest}
\title{Perform statistical test to check for DE between final stages of every
 trajectory.}
\usage{
diffEndTest(models, omnibus = TRUE, pairwise = FALSE)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{omnibus}{If TRUE, test for all pairwise comparisons simultaneously.}

\item{pairwise}{If TRUE, test for all pairwise comparisons independently.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed.
}
\description{
Perform statistical test to check for DE between final stages of every
 trajectory.
}
\examples{
data(gamList, package = "tradeSeq")
diffEndTest(gamList, omnibus = TRUE, pairwise = TRUE)
}
