% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffEndTest.R
\docType{methods}
\name{diffEndTest}
\alias{diffEndTest}
\alias{diffEndTest,SingleCellExperiment-method}
\alias{diffEndTest,list-method}
\title{Perform statistical test to check for DE between final stages of every
 lineage.}
\usage{
diffEndTest(models, ...)

\S4method{diffEndTest}{SingleCellExperiment}(models, global = TRUE,
  pairwise = FALSE)

\S4method{diffEndTest}{list}(models, global = TRUE, pairwise = FALSE)
}
\arguments{
\item{models}{Typically the output from
\code{\link{fitGAM}}, either a list of fitted GAM models, or an object of
\code{SingleCellExperiment} class.}

\item{global}{If TRUE, test for all pairwise comparisons simultaneously.}

\item{pairwise}{If TRUE, test for all pairwise comparisons independently.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed. If the testing
 procedure was unsuccessful, the procedure will return NA test statistics and
 p-values.
}
\description{
Perform statistical test to check for DE between final stages of every
 lineage.
}
\examples{
data(gamList, package = "tradeSeq")
diffEndTest(gamList, global = TRUE, pairwise = TRUE)
}
