% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/earlyDETest.R
\docType{methods}
\name{earlyDETest}
\alias{earlyDETest}
\alias{earlyDETest,SingleCellExperiment-method}
\alias{earlyDETest,list-method}
\title{Perform test of early differences between lineages}
\usage{
earlyDETest(models, ...)

\S4method{earlyDETest}{SingleCellExperiment}(models, global = TRUE,
  pairwise = FALSE, knots = NULL, nPoints = 100)

\S4method{earlyDETest}{list}(models, global = TRUE, pairwise = FALSE,
  knots = NULL, nPoints = 100)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{...}{parameters including:}

\item{global}{If TRUE, test for all pairwise comparisons simultaneously.}

\item{pairwise}{If TRUE, test for all pairwise comparisons independently.}

\item{knots}{A vector of length 2 specifying the knots before and after the
branching of interest.}

\item{nPoints}{the number of points to be compared between lineages.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed.
}
\description{
Perform test of early differences between lineages
}
\details{
To help in choosing the knots, the \code{\link{plotGeneCount}}
 function has a models optional parameter that can be used to visualize
  where the knots are. This helps the user to decide which knots to use when
   defining the branching
}
\examples{
data(gamList, package = "tradeSeq")
earlyDETest(gamList, knots = c(1, 2), global = TRUE, pairwise = TRUE)
}
