% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{earlyDETest}
\alias{earlyDETest}
\title{Perform test of early differences between lineages}
\usage{
earlyDETest(models, output = "both")
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{output}{The type of output. Can be "pval", "wald" or "both". Default to "both".}
}
\description{
Perform test of early differences between lineages
}
