% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{endPointTest}
\alias{endPointTest}
\title{Perform statistical test to check for DE between final stages of every trajectory.}
\usage{
endPointTest(models, omnibus = TRUE, pairwise = FALSE, ...)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform statistical test to check for DE between final stages of every trajectory.
}
