% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/evaluateK.R
\name{evaluateK}
\alias{evaluateK}
\alias{evaluateK,matrix-method}
\alias{evaluateK,dgCMatrix-method}
\alias{evaluateK,SingleCellExperiment-method}
\alias{evaluateK,CellDataSet-method}
\title{Evaluate the optimal number of knots required for fitGAM.}
\usage{
evaluateK(counts, ...)

\S4method{evaluateK}{matrix}(
  counts,
  k = 3:10,
  nGenes = 500,
  sds = NULL,
  pseudotime = NULL,
  cellWeights = NULL,
  plot = TRUE,
  U = NULL,
  weights = NULL,
  offset = NULL,
  aicDiff = 2,
  verbose = TRUE,
  control = mgcv::gam.control(),
  sce = FALSE,
  family = "nb",
  ...
)

\S4method{evaluateK}{dgCMatrix}(
  counts,
  k = 3:10,
  nGenes = 500,
  sds = NULL,
  pseudotime = NULL,
  cellWeights = NULL,
  plot = TRUE,
  U = NULL,
  weights = NULL,
  offset = NULL,
  aicDiff = 2,
  verbose = TRUE,
  control = mgcv::gam.control(),
  sce = FALSE,
  family = "nb",
  ...
)

\S4method{evaluateK}{SingleCellExperiment}(
  counts,
  k = 3:10,
  nGenes = 500,
  sds = NULL,
  pseudotime = NULL,
  cellWeights = NULL,
  plot = TRUE,
  U = NULL,
  weights = NULL,
  offset = NULL,
  aicDiff = 2,
  verbose = TRUE,
  control = mgcv::gam.control(),
  sce = FALSE,
  family = "nb",
  ...
)

\S4method{evaluateK}{CellDataSet}(
  counts,
  k = 3:10,
  nGenes = 500,
  sds = NULL,
  pseudotime = NULL,
  cellWeights = NULL,
  plot = TRUE,
  U = NULL,
  weights = NULL,
  offset = NULL,
  aicDiff = 2,
  verbose = TRUE,
  control = mgcv::gam.control(),
  sce = FALSE,
  family = "nb",
  ...
)
}
\arguments{
\item{counts}{The count matrix, genes in rows and cells in columns.}

\item{...}{parameters including:}

\item{k}{The range of knots to evaluate. `3:10` by default.}

\item{nGenes}{The number of genes to use in the evaluation. Genes will be
randomly selected. 500 by default.}

\item{sds}{Slingshot object containing the lineages.}

\item{pseudotime}{a matrix of pseudotime values, each row represents a cell
and each column represents a lineage.}

\item{cellWeights}{a matrix of cell weights defining the probability that a
cell belongs to a particular lineage. Each row represents a cell and each
column represents a lineage.}

\item{plot}{Whether to display diagnostic plots. Default to \code{TRUE}.}

\item{U}{The design matrix of fixed effects. The design matrix should not
contain an intercept to ensure identifiability.}

\item{weights}{Optional: a matrix of weights with identical dimensions
as the \code{counts} matrix. Usually a matrix of zero-inflation weights.}

\item{offset}{Optional: the offset, on log-scale. If NULL, TMM is used to
account for differences in sequencing depth, see \code{fitGAM}.}

\item{aicDiff}{Used for selecting genes with significantly varying AIC values
over the range of evaluated knots to make the barplot output. Default is set
to 2, meaning that only genes whose AIC range is larger than 2 will be used
to check for the optimal number of knots through the barplot visualization
that is part of the output of this function.}

\item{verbose}{logical, should progress be verbose?}

\item{control}{Control object for GAM fitting, see \code{mgcv::gam.control()}.}

\item{sce}{Logical, should a \code{SingleCellExperiment} object be returned?}

\item{family}{The distribution assumed, currently only \code{"nb"}
(negative binomial) is supported.}
}
\value{
A plot of average AIC value over the range of selected knots, and a
matrix of AIC values for the selected genes (rows) and the range of knots
(columns).
}
\description{
Evaluate the optimal number of knots required for fitGAM.

Evaluate an appropriate number of knots.
}
\examples{
## This is an artifical example, please check the vignette for a realistic one.
set.seed(8)
data(sds, package="tradeSeq")
loadings <- matrix(runif(2000*2,-2,2), nrow=2, ncol=2000)
counts <- round(abs(t(slingshot::reducedDim(sds) \%*\% loadings)))+100
aicK <- evaluateK(counts = counts, sds=sds,
                  nGenes=100, k=3:5, verbose=FALSE)
}
