% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_monocle_info}
\alias{extract_monocle_info}
\title{Extract info from Monocle models}
\usage{
extract_monocle_info(cds)
}
\arguments{
\item{cds}{A \code{\link{CellDataSet}}}
}
\value{

}
\description{
This function extracts info that will be used downstream to make
\code{\link{CellDataSet}} objects compatible with a \code{\lin{tradeSeq}}
analysis
}
\details{
For now, this only works for the DDRTree dimentionality reduction. 
It is the one recommanded by the Monocle developers.
}
