% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fitGAM}
\alias{fitGAM}
\title{Fit GAM model}
\usage{
fitGAM(counts, U = NULL, pseudotime, cellWeights, weights = NULL,
  seed = 81, offset = NULL, verbose = FALSE, nknots = 10)
}
\arguments{
\item{counts}{the count matrix.}

\item{U}{the design matrix of fixed effects. The design matrix should not
contain an intercept to ensure identifiability.}

\item{pseudotime}{a matrix of pseudotime values, each row represents a cell
and each column represents a lineage.}

\item{cellWeights}{a matrix of cell weights defining the probability that a
cell belongs to a particular lineage. Each row represents a cell and each
column represents a lineage.}

\item{weights}{a matrix of weights with identical dimensions
as the \code{counts} matrix. Usually a matrix of zero-inflation weights.}

\item{seed}{the seed used for assigning cells to trajectories}

\item{offset}{the offset, on log-scale, to account for differences in
sequencing depth.}

\item{verbose}{Whether to display the progress bar or not.}

\item{nknots}{Number of knots used to fit the GAM.}
}
\value{
A list of length the number of genes
 (number of rows of \code{counts}). Each element of the list is either a
  \code{\link{gamObject}} if the fiting procedure converged, or an error
   message.
}
\description{
Fit GAM model
}
\examples{
set.seed(8)
data(se, package = "tradeSeq")
se <- se[( 20:31)[-7], 25:40]
pseudotimes <- matrix(runif(ncol(se) * 2, 0, 5), ncol = 2)
cellWeights <- matrix(runif(ncol(se) * 2, 0, 1), ncol = 2)
gamList <- fitGAM(counts = as.matrix(
                      SummarizedExperiment::assays(se)$counts),
                  pseudotime = pseudotimes, cellWeights = cellWeights,
                  nknots = 5, verbose = TRUE)
gamList[[1]]
}
