% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fitGAM}
\alias{fitGAM}
\title{fitGAM}
\usage{
fitGAM(counts, U = NULL, pseudotime, cellWeights, weights = NULL,
  seed = 81, offset = NULL, nknots = 6, verbose = TRUE,
  parallel = FALSE, BPPARAM = BiocParallel::bpparam(),
  control = mgcv::gam.control())
}
\arguments{
\item{counts}{the count matrix.}

\item{U}{the design matrix of fixed effects. The design matrix should not
contain an intercept to ensure identifiability.}

\item{pseudotime}{a matrix of pseudotime values, each row represents a cell
and each column represents a lineage.}

\item{cellWeights}{a matrix of cell weights defining the probability that a
cell belongs to a particular lineage. Each row represents a cell and each
column represents a lineage.}

\item{weights}{a matrix of weights with identical dimensions
as the \code{counts} matrix. Usually a matrix of zero-inflation weights.}

\item{seed}{the seed used for assigning cells to lineages.}

\item{offset}{the offset, on log-scale. If NULL, TMM is used to account for
differences in sequencing depth., see \code{edgeR::calcNormFactors}.
Alternatively, this may also be a matrix of the same dimensions as the
expression matrix.}

\item{nknots}{Number of knots used to fit the GAM. Defaults to 6.}

\item{verbose}{Logical, should progress be printed?}

\item{parallel}{Logical, defaults to FALSE. Set to TRUE if you want to
parallellize the fitting.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{\link[BiocParallel]{bpparam}} for details.}

\item{control}{Variables to control fitting of the GAM, see
\code{gam.control}.}
}
\value{
A list of length the number of genes
 (number of rows of \code{counts}). Each element of the list is either a
  \code{\link{gamObject}} if the fiting procedure converged, or an error
   message.
}
\description{
fitGAM
}
\examples{
set.seed(8)
download.file("https://github.com/statOmics/tradeSeqPaper/raw/master/data/se_paul.rda",
destfile="./se_paul.rda")
load("./se_paul.rda")
se <- se[( 20:31)[-7], 25:40]
pseudotimes <- matrix(runif(ncol(se) * 2, 0, 5), ncol = 2)
cellWeights <- matrix(runif(ncol(se) * 2, 0, 1), ncol = 2)
gamList <- fitGAM(counts = as.matrix(
                      SummarizedExperiment::assays(se)$counts),
                  pseudotime = pseudotimes, cellWeights = cellWeights,
                  nknots = 5)
}
