% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fitGAM.R
\docType{methods}
\name{fitGAM}
\alias{fitGAM}
\alias{fitGAM,matrix-method}
\title{fitGAM}
\usage{
fitGAM(counts, ...)

\S4method{fitGAM}{matrix}(counts, sds = NULL, pseudotime = NULL,
  cellWeights = NULL, U = NULL, weights = NULL, offset = NULL,
  nknots = 6, verbose = TRUE, parallel = FALSE,
  BPPARAM = BiocParallel::bpparam(), control = mgcv::gam.control(),
  sce = FALSE, family = "nb")
}
\arguments{
\item{counts}{the count matrix.}

\item{...}{parameters including:}

\item{sds}{an object of class \code{SlingshotDataSet}, typically obtained
after running Slingshot. If this is provided, \code{pseudotime} and
\code{cellWeights} arguments are derived from this object.}

\item{pseudotime}{a matrix of pseudotime values, each row represents a cell
and each column represents a lineage.}

\item{cellWeights}{a matrix of cell weights defining the probability that a
cell belongs to a particular lineage. Each row represents a cell and each
column represents a lineage.}

\item{U}{the design matrix of fixed effects. The design matrix should not
contain an intercept to ensure identifiability.}

\item{weights}{a matrix of weights with identical dimensions
as the \code{counts} matrix. Usually a matrix of zero-inflation weights.}

\item{offset}{the offset, on log-scale. If NULL, TMM is used to account for
differences in sequencing depth., see \code{edgeR::calcNormFactors}.
Alternatively, this may also be a matrix of the same dimensions as the
expression matrix.}

\item{nknots}{Number of knots used to fit the GAM. Defaults to 6.}

\item{verbose}{Logical, should progress be printed?}

\item{parallel}{Logical, defaults to FALSE. Set to TRUE if you want to
parallellize the fitting.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{bpparam} in \code{BiocParallel} package for details.}

\item{control}{Variables to control fitting of the GAM, see
\code{gam.control}.}

\item{sce}{Should output be of SingleCellExperiment class? This argument
should not be changed by users.}

\item{family}{The assumed distribution for the response, set to \code{"nb"}
by default.}
}
\value{
A list of length the number of genes
 (number of rows of \code{counts}). Each element of the list is either a
  \code{\link{gamObject}} if the fiting procedure converged, or an error
   message.
}
\description{
This fits the NB-GAM model as described in Van den Berge et al.[2019]
}
\examples{
set.seed(8)
data(crv, package="tradeSeq")
data(countMatrix, package="tradeSeq")
gamList <- fitGAM(counts = as.matrix(countMatrix),
                  sds = crv,
                  nknots = 5)
}
