% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getSmootherPvalues}
\alias{getSmootherPvalues}
\title{Get smoother p-value}
\usage{
getSmootherPvalues(models)
}
\arguments{
\item{models}{the GAM models, typically the output from \code{\link{fitGAM}}.}
}
\value{
a matrix with the p-value associated with each lineage's smoother.
 The matrix has one row per gene where the fitting procedure converged.
}
\description{
Get smoother p-value
}
\examples{
data(gamList, package = "tradeR")
getSmootherPvalues(gamList)
}
