% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getSmootherPvalues}
\alias{getSmootherPvalues}
\title{Get smoother p-value}
\usage{
getSmootherPvalues(models)
}
\arguments{
\item{models}{the GAM models, typically the output from \code{\link{fitGAM}}.}
}
\description{
Get smoother p-value
}
