% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmootherTestStats.R
\name{getSmootherTestStats}
\alias{getSmootherTestStats}
\title{Get smoother Chi-squared test statistics}
\usage{
getSmootherTestStats(models)
}
\arguments{
\item{models}{the GAM models, typically the output from \code{\link{fitGAM}}.
Note that this function only works when \code{models} is a list.}
}
\value{
a matrix with the wald statistics associated with each lineage's
 smoother. The matrix has one row per gene where the fitting procedure
  converged.
}
\description{
Get smoother Chi-squared test statistics
}
\examples{
data(gamList, package = "tradeSeq")
getSmootherPvalues(gamList)
}
