% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{identicalTest}
\alias{identicalTest}
\title{Perform test between lineages to check whether the gene dynamics are identical.}
\usage{
identicalTest(models)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}
}
\description{
Perform test between lineages to check whether the gene dynamics are identical.
}
