% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{patternTest}
\alias{patternTest}
\title{Perform pattern test between lineages}
\usage{
patternTest(models, nPoints = 100, omnibus = TRUE, pairwise = FALSE,
  ...)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{nPoints}{the numboer of points to be compared between lineages.}
}
\description{
Perform pattern test between lineages
}
