% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternTest.R
\docType{methods}
\name{patternTest}
\alias{patternTest}
\alias{patternTest,SingleCellExperiment-method}
\title{Assess differential expression pattern between lineages.}
\usage{
\S4method{patternTest}{list}(models, global = TRUE, pairwise = FALSE,
  nPoints = 100)

\S4method{patternTest}{SingleCellExperiment}(models, global = TRUE,
  pairwise = FALSE, nPoints = 100)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{global}{If TRUE, test for all pairwise comparisons simultaneously.}

\item{pairwise}{If TRUE, test for all pairwise comparisons independently.}

\item{nPoints}{the number of points to be compared between lineages.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed. If the testing
 procedure was unsuccessful, the procedure will return NA test statistics and
 p-values.
}
\examples{
data(gamList, package = "tradeSeq")
patternTest(gamList, global = TRUE, pairwise = TRUE)
}
