% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotGeneCount}
\alias{plotGeneCount}
\title{Plot the gene in reduced dimension space}
\usage{
plotGeneCount(curve, counts, gene = NULL, clusters = NULL, models = NULL,
  title = NULL)
}
\arguments{
\item{curve}{The output from a lineage computation}

\item{counts}{the count matrix.}

\item{gene}{The name of gene for which you want to plot the count or the row
number of that gene in the count matrix. Alternatively, one can specify
the cluster arguments}

\item{clusters}{The assignation of each cell to a cluster. Used to color the
plot. Either \code{clusters} or \code{gene} must be supplied.}

\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}. Used to display the knots.}

\item{title}{Title for the plot.}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot the gene in reduced dimension space
}
\details{
If both \code{gene} and \code{clusters} arguments are supplied, the
 plot will be colored according to gene count level.
}
\examples{
download.file("https://github.com/statOmics/tradeSeqPaper/raw/master/data/se_paul.rda",
destfile="./se_paul.rda")
load("./se_paul.rda")
set.seed(97)
data(se, package = "tradeSeq")
rd <- SingleCellExperiment::reducedDims(se)$UMAP
cl <- kmeans(rd, centers = 7)$cluster
library(slingshot)
lin <- getLineages(rd, clusterLabels = cl, start.clus = 4)
crv <- getCurves(lin)
counts <- as.matrix(SummarizedExperiment::assays(se)$counts)
filt <- rowSums(counts > 8) > ncol(counts)/100
counts <- counts[filt, ]
gamList <- fitGAM(counts = counts,
 pseudotime = slingPseudotime(crv, na = FALSE),
 cellWeights = slingCurveWeights(crv))
plotGeneCount(crv, counts, gene = "Mpo")
}
