% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSmoothers}
\alias{plotSmoothers}
\title{plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.}
\usage{
plotSmoothers(m, nPoints = 100, lwd = 2, size = 2/3,
  xlab = "pseudotime", ylab = " expression + 1 (log-scale)")
}
\arguments{
\item{m}{the fitted model of a given gene}

\item{nPoints}{The number of points used to extraplolate the fit}

\item{lwd}{Line width of the smoother. Passed to \code{\link{geom_line}}}

\item{size}{Character expansion of the data points. Passed to \code{\link{geom_point}}}

\item{xlab}{x-axis label. Passed to \code{\link{labs}}}

\item{ylab}{y-axis label. Passed to \code{\link{labs}}}
}
\value{
A \code{\link{ggplot}} object
}
\description{
plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.
}
\examples{
data(gamList, package = "tradeSeq")
plotSmoothers(gamList[[4]])
}
