% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSmoothers}
\alias{plotSmoothers}
\title{plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.}
\usage{
plotSmoothers(m, nPoints = 100, lwd = 2, cex = 2/3, pch = 16,
  xlab = "pseudotime", ylab = " expression + 1 (log-scale)", ...)
}
\arguments{
\item{m}{the fitted model of a given gene}

\item{nPoints}{The number of points used to extraplolate the fit}

\item{lwd}{Line width of the smoother. Passed to \code{\link{lines}}}

\item{cex}{Character expansion of the data points. Passed to \code{\link{plot}}}

\item{pch}{Plotting character of the data points. Passed to \code{\link{plot}}}

\item{xlab}{x-axis label. Passed to \code{\link{plot}}}

\item{ylab}{y-axis label. Passed to \code{\link{plot}}}

\item{...}{Further arguments passed to \code{\link{plot}}}
}
\value{
A plot that is printed.
}
\description{
plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.
}
\examples{
data(gamList, package = "tradeSeq")
plotSmoothers(gamList[[4]])
}
