% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSmoothers.R
\name{plotSmoothers}
\alias{plotSmoothers}
\alias{plotSmoothers,gam-method}
\alias{plotSmoothers,SingleCellExperiment-method}
\title{Plot the log-transformed counts and the fitted values for a particular
 gene along all lineages}
\usage{
plotSmoothers(models, ...)

\S4method{plotSmoothers}{gam}(
  models,
  nPoints = 100,
  lwd = 2,
  size = 2/3,
  xlab = "Pseudotime",
  ylab = "Log(expression + 1)",
  border = TRUE,
  alpha = 1
)

\S4method{plotSmoothers}{SingleCellExperiment}(
  models,
  counts,
  gene,
  nPoints = 100,
  lwd = 2,
  size = 2/3,
  xlab = "Pseudotime",
  ylab = "Log(expression + 1)",
  border = TRUE,
  alpha = 1
)
}
\arguments{
\item{models}{Either the \code{SingleCellExperiment} object obtained after
running \code{fitGAM}, or the specific GAM model for the corresponding gene,
if working with the list output of \code{tradeSeq}.}

\item{...}{parameters including:}

\item{nPoints}{The number of points used to extraplolate the fit.
Defaults to 100.}

\item{lwd}{Line width of the smoother. Passed to \code{\link{geom_line}}.}

\item{size}{Character expansion of the data points. Passed to \code{\link{geom_point}}.}

\item{xlab}{x-axis label. Passed to \code{\link{labs}}.}

\item{ylab}{y-axis label. Passed to \code{\link{labs}}.}

\item{border}{Logical: should a white border be drawn around the mean smoother.}

\item{alpha}{Numeric between 0 and 1, determines the transparancy of data points,
see \code{scale_color_viridis_d}.}

\item{counts}{The matrix of gene expression counts.}

\item{gene}{Gene name or row in count matrix of gene to plot.}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot the smoothers estimated by \code{tradeSeq}.
}
\examples{
data(gamList, package = "tradeSeq")
plotSmoothers(gamList[[4]])
}
