% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotSmoothers}
\alias{plotSmoothers}
\title{plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.}
\usage{
plotSmoothers(m, nPoints = 100, lwd = 2, ...)
}
\arguments{
\item{m}{the fitted model of a given gene}

\item{nPoints}{The number of points used to extraplolate the fit}

\item{...}{Further arguments passed to \code{\link{plot}}}
}
\value{
A plot that is printed.
}
\description{
plot the logged-transformed counts and the fitted values for a particular
 gene along all trajectories.
}
\examples{
data(gamList, package = "tradeSeq")
plotSmoothers(gamList[[4]])
}
