% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSmoothers.R
\docType{methods}
\name{plotSmoothers}
\alias{plotSmoothers}
\alias{plotSmoothers,gam-method}
\alias{plotSmoothers,SingleCellExperiment-method}
\title{Plot the log-transformed counts and the fitted values for a particular
 gene along all lineages}
\usage{
plotSmoothers(models, ...)

\S4method{plotSmoothers}{gam}(models, nPoints = 100, lwd = 2,
  size = 2/3, xlab = "Pseudotime", ylab = "Log(expression + 1)")

\S4method{plotSmoothers}{SingleCellExperiment}(models, counts, gene,
  nPoints = 100, lwd = 2, size = 2/3, xlab = "Pseudotime",
  ylab = "Log(expression + 1)")
}
\arguments{
\item{...}{parameters including:}

\item{nPoints}{The number of points used to extraplolate the fit}

\item{lwd}{Line width of the smoother. Passed to \code{\link{geom_line}}}

\item{size}{Character expansion of the data points. Passed to \code{\link{geom_point}}}

\item{xlab}{x-axis label. Passed to \code{\link{labs}}}

\item{ylab}{y-axis label. Passed to \code{\link{labs}}}

\item{m}{the fitted model of a given gene}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot the log-transformed counts and the fitted values for a particular
 gene along all lineages
}
\examples{
data(gamList, package = "tradeSeq")
plotSmoothers(gamList[[4]])
}
