% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predictSmooth.R
\name{predictSmooth}
\alias{predictSmooth}
\alias{predictSmooth,SingleCellExperiment-method}
\title{predictSmooth}
\usage{
predictSmooth(models, ...)

\S4method{predictSmooth}{SingleCellExperiment}(models, counts, gene, nPoints = 100)
}
\arguments{
\item{models}{Either the \code{SingleCellExperiment} object obtained after
running \code{fitGAM}, or the specific GAM model for the corresponding gene,
if working with the list output of \code{tradeSeq}.}

\item{...}{parameters including:}

\item{counts}{The matrix of gene expression counts.}

\item{gene}{Either a vector of gene names or an integer vector, corresponding
to the rows of the genes.}

\item{nPoints}{The number of points used to create the grid along the
smoother for each lineage.}
}
\value{
A \code{matrix} with estimated averages.
}
\description{
Get estimated smoothers estimated by \code{tradeSeq} along a
grid. This function does not return fitted values but rather the predicted
mean smoother, for a grid of user-defined points.
}
