% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/predictSmooth.R
\docType{methods}
\name{predictSmooth}
\alias{predictSmooth}
\alias{predictSmooth,SingleCellExperiment-method}
\title{predictSmooth}
\usage{
predictSmooth(models, ...)

\S4method{predictSmooth}{SingleCellExperiment}(models, counts, gene,
  nPoints = 100)
}
\arguments{
\item{models}{Either the \code{SingleCellExperiment} object obtained after
running \code{fitGAM}, or the specific GAM model for the corresponding gene,
if working with the list output of \code{tradeSeq}.}

\item{...}{parameters including:}

\item{counts}{The matrix of gene expression counts.}

\item{gene}{Gene name of gene to plot.}

\item{nPoints}{The number of points used to extraplolate the fit}
}
\value{
A \code{matrix} with fitted values.
}
\description{
Get fitted values from the smoothers estimated by \code{tradeSeq}.
}
