% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startVsEndTest.R
\docType{methods}
\name{startVsEndTest}
\alias{startVsEndTest}
\alias{startVsEndTest,list-method}
\title{Perform statistical test to check for DE between starting point and the end
stages of every lineage}
\usage{
\S4method{startVsEndTest}{SingleCellExperiment}(models, global = TRUE,
  lineages = FALSE, pseudotimeValues = NULL)

\S4method{startVsEndTest}{list}(models, global = TRUE,
  lineages = FALSE, pseudotimeValues = NULL)
}
\arguments{
\item{models}{the list of GAMs, typically the output from
\code{\link{fitGAM}}.}

\item{global}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}

\item{pseudotimeValues}{a vector of length 2, specifying two pseudotime
values to be compared against each other, for every lineage of
 the trajectory.
 @details Note that this test assumes that all lineages start at a
 pseudotime value of zero, which is the starting point against which the
 end point is compared.}
}
\value{
A matrix with the wald statistic, the number of df and the p-value
 associated with each gene for all the tests performed. If the testing
 procedure was unsuccessful, the procedure will return NA test statistics and
 p-values. If both \code{global} and \code{lineages} are TRUE, then a matrix
 of p-values is returned.
}
\description{
Perform statistical test to check for DE between starting point and the end
stages of every lineage
}
\examples{
data(gamList, package = "tradeSeq")
startVsEndTest(gamList, global = TRUE, lineages = TRUE)
}
